+++ dev-incubator-2/dev-incubator-2/VehicleOld.cs
@@ -10,12 +10,12 @@ namespace dev_incubator_2
     { 
         readonly int age;
         const int startOldAge = 5;
-        const double surchargeOld = 0.6d;
+        const decimal surchargeOld = 0.6M;
 
         public YearsOperation endOperation { get; set; }
         public VehicleOld() { }
 
-        public VehicleOld(string name, double averageCostKilometer, double averageKilometers, YearsOperation endOperation, int age)
+        public VehicleOld(string name, decimal averageCostKilometer, decimal averageKilometers, YearsOperation endOperation, int age)
         {
             vehicleCosts = new VehicleCosts(name, averageCostKilometer);
             this.averageKilometers = averageKilometers;
@@ -23,9 +23,9 @@ namespace dev_incubator_2
             this.age = age;
         }       
 
-        public override double GetTotalCosts() 
+        public override decimal GetTotalCosts() 
         {
-            double surcharge = 0;
+            decimal surcharge = 0;
             if (age > startOldAge) surcharge = surchargeOld;
             return (vehicleCosts.averageCostKilometer + surcharge) * averageKilometers;
         }
